requirements.txt
google-adk==1.14.0
langchain-community==0.3.27
wikipedia==1.4.0
python-dotenv==1.0.0
google-cloud-firestore>=2.16.0
google-cloud-logging>=3.10.0
mcp>=1.5.0
